Adding a new section in trouble-shooting based on a recent set of errors new users have been encountering. This issue is not a gkyl issue per se, but rather something to do with some operating systems not setting the default language for string/character parsing and thus some clusters (such as Stampede 2 or Frontera) getting confused about what the settings should be when you log in.
If this language is not set, users reported esoteric errors during the build and configure process (like not finding the C compiler). I am not sure if this is unique to older version of Ubuntu (the help on the internet suggested this was commonly the case). Suffice to say so long as the language (where I chose the default recommendation to be en_US.UTS-8) is set, the warnings at login go away and users are able to build and run jobs on the cluster.

I have also clarified that the installation using machine files is recommended when installing from source based on some confusion new users were having about what they should do after git cloning gkyl (with some new users trying to install luajit themselves without realizing using a machine file is part of our recommended guide when installing from source).